Move codable FeedStore to its own file in production
