+++ EssentialFeed/EssentialFeed/Feed Cache/CodableFeedStore.swift
@@ -1,0 +1,83 @@
+//
+//  CodableFeedStore.swift
+//  EssentialFeed
+//
+//  Created by Jules on 24/02/2025.
+//
+
+import Foundation
+
+public class CodableFeedStore: FeedStore {
+    private struct Cache: Codable {
+        let feed: [CodableFeedImage]
+        let timestamp: Date
+        
+        var localFeed: [LocalFeedImage] {
+            return feed.map { $0.local }
+        }
+    }
+    
+    private struct CodableFeedImage: Codable {
+        private let id: UUID
+        private let description: String?
+        private let location: String?
+        private let url: URL
+        
+        init(_ image: LocalFeedImage) {
+            self.id = image.id
+            self.description = image.description
+            self.location = image.location
+            self.url = image.url
+        }
+        
+        var local : LocalFeedImage {
+            return LocalFeedImage(id: id, description: description, location: location, url: url)
+        }
+    }
+    
+    private let storeURL: URL
+    
+    public  init(_ storeURL: URL) {
+        self.storeURL = storeURL
+    }
+    
+    public func retrieve(completion: @escaping RetrievalCompletion) {
+        
+        guard let data = try? Data(contentsOf: storeURL) else {
+            return completion(.empty)
+        }
+        do {
+            let decoder = JSONDecoder()
+            let cache = try decoder.decode(Cache.self, from: data)
+            completion(.found(feed: cache.localFeed, timestamp: cache.timestamp))
+        } catch {
+            completion(.failure(error))
+        }
+        
+    }
+    
+    public func insert(_ feed: [LocalFeedImage], with timestamp: Date, completion: @escaping InsertionCompletion)
+    {
+        do {
+            let encoder = JSONEncoder()
+            let cache = Cache(feed: feed.map(CodableFeedImage.init), timestamp: timestamp)
+            let encoded = try encoder.encode(cache)
+            try encoded.write(to: storeURL)
+            completion(nil)
+        } catch {
+            completion(error)
+        }
+    }
+    
+    public func deleteCachedFeed(completion: @escaping DeletionCompletion) {
+        guard FileManager.default.fileExists(atPath: storeURL.path) else {
+            return completion(nil)
+        }
+        do {
+            try FileManager.default.removeItem(at: storeURL)
+            completion(nil)
+        } catch {
+            completion(error)
+        }
+    }
+}
